+++ utils/nodes-csv.sh
@@ -21,12 +21,20 @@ echo -e "\
 </osm-script> \
 " >/tmp/nodes_csv_req
 
+REQUEST_METHOD=
+/home/roland/osm-3s/cgi-bin/interpreter </tmp/nodes_csv_req >/tmp/nodes_csv_result.1
+RESPONSE_TYPE=`head -n 1 </tmp/nodes_csv_result.1`
+if [[ $RESPONSE_TYPE != "Content-type: application/osm3s" ]]; then
+{
+  cat </tmp/nodes_csv_result.1
+  exit 0
+};
+fi
+dd if=/tmp/nodes_csv_result.1 of=/tmp/nodes_csv_result.2 bs=1 skip=56
+
 echo "Content-Type: text/html; charset=utf-8"
 echo
 
-REQUEST_METHOD=
-/home/roland/osm-3s/cgi-bin/interpreter </tmp/nodes_csv_req >/tmp/nodes_csv_result.1
-dd if=/tmp/nodes_csv_result.1 of=/tmp/nodes_csv_result.2 bs=1 skip=56
 if [[ $ARG_2 == "forward" ]]; then
 {
   gunzip </tmp/nodes_csv_result.2 | ../bin/nodes-csv-lat-lon-name --forward
